This is updated
Includes a refresh button that can account for turn and river cards. Also has a filter for flop and playerhand cards during initialization of the ArrayList. Needs to have for loops put in every scoring function that can count how many times a possibility is considered true. Also could use a better GUI.