I did everything except for the calculator itself
GUI is fully functional (besides refresh button), and actionListeners are working while inserting new values into the Cards. 0 is the new value of null for the turn and river cards, make sure to test that a turn or river card has both values !=0 before using them. Also added mutators to cards and hands, along with (mislabeled) mutators for each card's values in the Calculator class. All that is left is the WinPossibilities function and the output onto the GUI (that includes refresh button required).